docker/container_api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
aiortc>=1.9.0
mss>=9.0.0
Pillow>=10.0.0
